WIP: Separated loaders from import_gtfs-script
