+++ gtfspy/__init__.py
@@ -1,0 +1,1 @@
+import gtfspy.util as util
