+++ gtfspy/import_loaders/__init__.py
@@ -1,0 +1,15 @@
+from gtfspy.import_loaders.agency_loader import AgencyLoader
+from gtfspy.import_loaders.calendar_dates_loader import CalendarDatesLoader
+from gtfspy.import_loaders.calendar_loader import CalendarLoader
+from gtfspy.import_loaders.day_loader import DayLoader
+from gtfspy.import_loaders.day_trips_materializer import DayTripsMaterializer
+from gtfspy.import_loaders.feed_info_loader import FeedInfoLoader
+from gtfspy.import_loaders.frequencies_loader import FrequenciesLoader
+from gtfspy.import_loaders.trip_loader import TripLoader
+from gtfspy.import_loaders.metadata_loader import MetadataLoader
+from gtfspy.import_loaders.route_loader import RouteLoader
+from gtfspy.import_loaders.shape_loader import ShapeLoader
+from gtfspy.import_loaders.stop_distances_loader import StopDistancesLoader
+from gtfspy.import_loaders.stop_loader import StopLoader
+from gtfspy.import_loaders.stop_times_loader import StopTimesLoader
+from gtfspy.import_loaders.transfer_loader import TransfersLoader
